Adjustments to unit tests
